po-ldl1v: simplify: consolidate duplicate callbackHandler test file
Move the two unique removeEventListener-cleanup tests from
callbackHandler.test.ts into telegram.callbackHandler.test.ts, then
delete the now-redundant file. callbackHandler.test.ts was added for
the listener-leak fix but duplicated 11 of the 13 tests already in the
comprehensive suite, leaving two tests (backoff abort-listener cleanup)
with no equivalent. Net: 11 fewer duplicate tests, zero coverage lost.
